Extracted response-raise-error-if-failure logic out from client.
